requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
future==0.18.2
h11==0.12.0
pydantic==1.8.2
python-whois==0.7.3
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
